Continued to implement complex pedal editing
Most PedalEditAction functionalities are now implemented.

+++ src/mfx/uitk/pg/Tools.h
@@ -88,10 +88,10 @@ public:
 
 	static void    set_param_text (const Model &model, const View &view, int width, int index, float val, int slot_id, PiType type, NText *param_name_ptr, NText &param_val, NText *param_unit_ptr, NText *fx_name_ptr, bool group_unit_val_flag);
 	static MsgHandlerInterface::EvtProp
-	               change_param (Model &model, const View &view, int slot_id, PiType type, int index, float step, int step_index, int dir);
-	static double  change_param (double val_nrm, const Model &model, const View &view, int slot_id, PiType type, int index, float step, int step_index, int dir);
-	static double  change_param (double val_nrm, const View &view, const piapi::PluginDescInterface &desc_pi, const doc::PluginSettings &settings, int index, float step, int step_index, int dir);
-	static double  change_param (double val_nrm, const piapi::PluginDescInterface &desc_pi, int index, float step, int dir);
+	               change_param (Model &model, const View &view, int slot_id, PiType type, int index, double step, int step_index, int dir);
+	static double  change_param (double val_nrm, const Model &model, const View &view, int slot_id, PiType type, int index, double step, int step_index, int dir);
+	static double  change_param (double val_nrm, const View &view, const piapi::PluginDescInterface &desc_pi, const doc::PluginSettings &settings, int index, double step, int step_index, int dir);
+	static double  change_param (double val_nrm, const piapi::PluginDescInterface &desc_pi, int index, double step, int dir);
 	static int     find_ctrl_index (const ControlSource &src, const std::vector <CtrlSrcNamed> &ctrl_list);
 	static std::string
 	               find_ctrl_name (const ControlSource &src, const std::vector <CtrlSrcNamed> &ctrl_list);
